update base class
